added progress bar and other improvements

+++ AR_downloader.py
@@ -1,4 +1,4 @@
-import requests
+import requests 
 from requests.adapters import HTTPAdapter, Retry
 import os
 import pandas as pd
@@ -6,6 +6,7 @@ import logging
 import time
 import datetime
 import hashlib
+from tqdm import tqdm
 
 if __name__ == '__main__':
     runs_dir = 'AR_crawler_runs'
@@ -19,10 +20,13 @@ if __name__ == '__main__':
     runs_dir_list.sort(reverse=True) # list sorted from recent to older dates
     run_name = runs_dir_list[0] # most recent log selected
 
-    print(f'Crawler run to download: {run_name}')
+    logging.info(f'Crawler run to download: {run_name}')
 
     with open(run_name, 'r') as run:
         run_df = pd.read_csv(run)
+
+    # Setup progress bar for the command line 
+    pbar = tqdm(total=len(run_df), unit='files')
 
     # setup the requests to make the get() method resilient to failiures 
     # (see https://stackoverflow.com/questions/15431044/can-i-set-max-retries-for-requests-request)
@@ -36,15 +40,28 @@ if __name__ == '__main__':
     download_status = []
     download_time = []
     for index, row in run_df.iterrows():
-        file_path = row.loc['directory']
+        file_path = row.loc['file_path']
         download_url = row.loc['file_url']
-        print(f'Downloading file: {file_path}')
-        print(f'url: {download_url}')
+        
+        logging.info(f'Downloading file: {file_path}')
+        logging.debug(f'url: {download_url}')
+        pbar.update(1) # update progress bar
+        pbar.set_description(f'Downloading: {file_path}')
+        
+        # check if the file has already been downloaded, and if yes, skip it
+        if os.path.isfile(file_path):
+            download_status.append('already_downloaded')
+            download_time.append(0)
+            checksums.append('')
+            continue
+
         os.makedirs(os.path.dirname(file_path), exist_ok=True)
         downloaded_file = s.get(download_url)
         download_status.append(downloaded_file.status_code)
         download_time.append(datetime.datetime.now())
-        if downloaded_file.status_code == 200:    # status code = 200 means the download went fine, anything else means something went wrong
+
+        # status code = 200 means the download went fine, anything else means something went wrong
+        if downloaded_file.status_code == 200:    
             checksums.append(hashlib.md5(downloaded_file.content).hexdigest())
             with open(file_path, 'wb') as f:
                 f.write(downloaded_file.content)
@@ -52,6 +69,9 @@ if __name__ == '__main__':
             checksums.append('')
         time.sleep(0.5) # waitting time to not overload the servers
     
+    # close progress bar
+    pbar.close() 
+    # Add columns to the datafram to save as a CSV
     run_df['download_md5'] = pd.Series(checksums)
     run_df['download_status'] = pd.Series(download_status)
     run_df['download_time'] = pd.Series(download_time)
